refactor: Waiting 헥사고날 아키텍처 패키지 구조로 수정
