+++ src/main/java/roomescape/waiting/application/dto/request/WaitingRequest.java
@@ -1,4 +1,4 @@
-package roomescape.waiting.dto.request;
+package roomescape.waiting.application.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import jakarta.validation.constraints.NotNull;
