+++ src/main/java/roomescape/waiting/domain/WaitingLine.java
@@ -1,4 +1,4 @@
-package roomescape.waiting;
+package roomescape.waiting.domain;
 
 import java.util.Comparator;
 import java.util.List;
