+++ src/test/java/roomescape/waiting/JdbcWaitingRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.waiting;
+package roomescape.waiting.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
-import roomescape.waiting.infrastructure.JdbcWaitingRepository;
+import roomescape.waiting.adapter.out.persistence.JdbcWaitingRepository;
 
 import java.time.LocalDate;
 import java.util.List;
